+++ R/import_data.R
@@ -16,8 +16,6 @@
 #'
 #' @export
 import_mhealth_csv <- function(filepath) {
-  tz_str <- "+0000"
-
   ncols <- readr::count_fields(filepath, readr::tokenizer_csv(), skip = 0, n_max = 1L)
   date_format <- readr::col_datetime(format = "%Y-%m-%d %H:%M:%OS")
   coltypes <- list(date_format)
@@ -46,6 +44,109 @@ import_mhealth_csv <- function(filepath) {
 
   return(df)
 }
+
+#' Import large raw multi-channel accelerometer data stored in mHealth Specification
+#' in chunks.
+#'
+#' \code{import_mhealth_csv_chunked} imports the raw multi-channel accelerometer
+#' data stored in mHealth Specification in chunks.
+#'
+#' @section How is it used in MIMS-unit algorithm?: This function is a File IO
+#'   function that is used to import data stored in mHealth Specification during
+#'   algorithm validation.
+#'
+#' @param filepath string. The filepath of the input data.
+#' @return list. The list contains three items. The first item is a generator
+#' function that each time it is called, it will
+#' return a list of `list(before_df, df, after_df)`. These are data.frames of
+#' imported chunks, ordered in time that can be loaded with
+#' the `MIMSunit::mims_unit` function directly. The second item is a
+#' `has_more_chunks` function used to check if all chunks are loaded.
+#' If it returns `FALSE`, it means the loading has ended. The third item is a
+#' `close` function which you can call at any moment to close the file loading.
+#' @family Filo I/O functions
+#'
+#' @export
+import_mhealth_csv_chunked <- function(filepath) {
+  options(digits.secs = 3)
+  chunk_size <- 1800 * 100
+  col_types <- c("character", "numeric", "numeric", "numeric")
+  col_names <- c("HEADER_TIME_STAMP", "X", "Y", "Z")
+
+  con <- file(filepath, open = "r")
+  g.env <- new.env()
+  g.env$before_df <- NULL
+  g.env$df <- NULL
+  g.env$after_df <- NULL
+
+
+  has_more_chunks <- function() {
+    is_open <- tryCatch(
+      {
+        isOpen(con)
+      },
+      error =
+        function(cond) {
+          return(FALSE)
+        }
+    )
+    if (is.null(g.env$after_df) || !is_open) {
+      return(FALSE)
+    }
+    if (nrow(g.env$after_df) == chunk_size) {
+      return(TRUE)
+    } else {
+      return(FALSE)
+    }
+  }
+
+  next_chunk <- function() {
+    g.env$before_df <- cbind(g.env$df)
+    if (!is.null(g.env$after_df)) {
+      g.env$df <- cbind(g.env$after_df)
+    } else {
+      df <- read.csv(
+        file = con,
+        header = TRUE,
+        skip = 0,
+        nrows = chunk_size,
+        colClasses = col_types,
+        col.names = col_names
+      )
+      df <- .convert_mhealth_timestamps(df)
+      g.env$df <- cbind(as.data.frame(df))
+    }
+
+    after_df <- read.csv(
+      file = con,
+      header = TRUE,
+      skip = 0,
+      nrows = chunk_size,
+      colClasses = col_types,
+      col.names = col_names
+    )
+    after_df <- .convert_mhealth_timestamps(after_df)
+    g.env$after_df <- cbind(as.data.frame(after_df))
+    if (nrow(after_df) == 0) {
+      g.env$after_df <- NULL
+    }
+
+    return(list(g.env$before_df, g.env$df, g.env$after_df))
+  }
+
+  close_connection <- function() {
+    close(con)
+  }
+  return(list(next_chunk, has_more_chunks, close_connection))
+}
+
+.convert_mhealth_timestamps <- function(dat) {
+  time_format <- "%Y-%m-%d %H:%M:%OS"
+  dat[["HEADER_TIME_STAMP"]] <-
+    strptime(x = dat[["HEADER_TIME_STAMP"]], format = time_format) + 5e-04
+  return(dat)
+}
+
 
 
 #' Import raw multi-channel accelerometer data stored in ActivPal3 csv format
@@ -107,6 +208,184 @@ import_activpal3_csv <- function(filepath, header = FALSE) {
   return(dat)
 }
 
+#' Import large raw multi-channel accelerometer data stored in Actigraph raw csv
+#' format in chunks
+#'
+#' \code{import_actigraph_csv_chunked} imports the raw multi-channel accelerometer data
+#' stored in Actigraph raw csv format. It supports files from the following
+#' devices: GT3X, GT3X+, GT3X+BT, GT9X, and GT9X-IMU.
+#'
+#' For old device (GT3X) that stores accelerometer values as digital voltage.
+#' The function will convert the values to \eqn{g} unit using the following
+#' equation.
+#'
+#' \deqn{x_g = \frac{x_{voltage}r}{(2 ^ r) - \frac{v}{2}}}
+#'
+#' Where \eqn{v} is the max voltage corresponding to the max accelerometer value
+#' that can be found in the meta section in the csv file; \eqn{r} is the
+#' resolution level which is the number of bits used to store the voltage
+#' values. \eqn{r} can also be found in the meta section in the csv file.
+#'
+#' @section How is it used in MIMS-unit algorithm?: This function is a File IO
+#'   function that is used to import data from Actigraph devices during
+#'   algorithm validation.
+#'
+#' @param filepath string. The filepath of the input data.
+#' @param in_voltage set as TRUE only when the input Actigraph csv file is in
+#'   analog quantized format and need to be converted into g value
+#' @param has_ts set as TRUE only when timestamp is provided as the first column
+#' @param header boolean. If TRUE, the input csv file will have column names in
+#'   the first row.
+#' @return list. The list contains three items. The first item is a generator
+#' function that each time it is called, it will
+#' return a list of `list(before_df, df, after_df)`. These are data.frames of
+#' imported chunks, ordered in time that can be loaded with
+#' the `MIMSunit::mims_unit` function directly. The second item is a
+#' `has_more_chunks` function used to check if all chunks are loaded.
+#' If it returns `FALSE`, it means the loading has ended. The third item is a
+#' `close` function which you can call at any moment to close the file loading.
+#'
+#' @family Filo I/O functions
+#' @export
+import_actigraph_csv_chunked <- function(filepath,
+                                         in_voltage = FALSE,
+                                         has_ts = TRUE,
+                                         header = TRUE) {
+  options(digits.secs = 3)
+  chunk_size <- 1800 * 100
+  actigraph_meta <- import_actigraph_meta(filepath)
+  if (has_ts) {
+    ncols <- 4
+    col_types <- c("character", "numeric", "numeric", "numeric")
+    col_names <- c("HEADER_TIME_STAMP", "X", "Y", "Z")
+  } else {
+    ncols <- 3
+    col_types <- c("numeric", "numeric", "numeric")
+    col_names <- c("X", "Y", "Z")
+  }
+  if (header) {
+    skip <- 11
+  } else {
+    skip <- 10
+  }
+  con <- file(filepath, "r")
+  g.env <- new.env()
+  g.env$before_dat <- NULL
+  g.env$dat <- NULL
+  g.env$after_dat <- NULL
+
+  has_more_chunks <- function() {
+    is_open <- tryCatch(
+      {
+        isOpen(con)
+      },
+      error =
+        function(cond) {
+          return(FALSE)
+        }
+    )
+    if (is.null(g.env$after_dat) || !is_open) {
+      return(FALSE)
+    }
+    if (nrow(g.env$after_dat) == chunk_size) {
+      return(TRUE)
+    } else {
+      return(FALSE)
+    }
+  }
+
+  next_chunk <- function() {
+    g.env$before_dat <- cbind(g.env$dat)
+    if (!is.null(g.env$after_dat)) {
+      g.env$dat <- cbind(g.env$after_dat)
+    } else {
+      dat <- read.csv(con,
+        header = FALSE,
+        skip = skip,
+        nrows = chunk_size,
+        colClasses = col_types,
+        col.names = col_names
+      )
+      if (!has_ts) {
+        dat <- .append_actigraph_timestamps(dat, actigraph_meta)
+      } else {
+        dat <- .convert_actigraph_timestamps(dat, actigraph_meta)
+      }
+      colnames(dat) <- c("HEADER_TIME_STAMP", "X", "Y", "Z")
+
+      if (in_voltage) {
+        dat <- .convert_voltage_to_g(dat, actigraph_meta)
+      }
+
+      g.env$dat <- as.data.frame(dat)
+    }
+    after_dat <- read.csv(con,
+      header = FALSE,
+      skip = 0,
+      nrows = chunk_size,
+      colClasses = col_types,
+      col.names = col_names
+    )
+    if (!has_ts) {
+      after_dat <- .append_actigraph_timestamps(after_dat, actigraph_meta)
+    } else {
+      after_dat <- .convert_actigraph_timestamps(after_dat, actigraph_meta)
+    }
+    colnames(after_dat) <- c("HEADER_TIME_STAMP", "X", "Y", "Z")
+
+    if (in_voltage) {
+      after_dat <- .convert_voltage_to_g(after_dat, actigraph_meta)
+    }
+
+    g.env$after_dat <- as.data.frame(after_dat)
+    if (nrow(after_dat) == 0) {
+      g.env$after_dat <- NULL
+    }
+    return(list(g.env$before_dat, g.env$dat, g.env$after_dat))
+  }
+
+  close_connection <- function() {
+    close(con)
+  }
+  return(list(next_chunk, has_more_chunks, close_connection))
+}
+
+.append_actigraph_timestamps <- function(dat, actigraph_meta) {
+  ts_col <-
+    seq(
+      from = actigraph_meta$st,
+      to = actigraph_meta$dt,
+      length.out = nrow(dat)
+    )
+  dat <- cbind(ts_col, dat)
+  return(dat)
+}
+
+.convert_actigraph_timestamps <- function(dat, actigraph_meta) {
+  time_format <-
+    ifelse(test = actigraph_meta$imu,
+      yes = "%Y-%m-%dT%H:%M:%OS",
+      no = "%m/%d/%Y %H:%M:%OS"
+    )
+  dat[["HEADER_TIME_STAMP"]] <-
+    strptime(x = dat[["HEADER_TIME_STAMP"]], format = time_format) + 5e-04
+  return(dat)
+}
+
+.convert_voltage_to_g <- function(dat, actigraph_meta) {
+  vs <- actigraph_meta$vs
+  res <- actigraph_meta$res
+
+  dat[, 2:ncol(dat)] <-
+    (dat[, 2:ncol(dat)] * vs / (2^res) - vs / 2) / (vs / res)
+  dat[, 2:ncol(dat)] <-
+    as.data.frame(apply(dat[, 2:ncol(dat)], 2, function(col) {
+      col[col == -5] <- 0
+      return(as.numeric(col))
+    }))
+  return(dat)
+}
+
 #' Import raw multi-channel accelerometer data stored in Actigraph raw csv
 #' format
 #'
@@ -148,18 +427,13 @@ import_actigraph_csv <-
            header = TRUE) {
     actigraph_meta <- import_actigraph_meta(filepath)
 
-    ncols <-
-      readr::count_fields(filepath,
-        readr::tokenizer_csv(),
-        n_max = 1,
-        skip = 11
-      )
     if (has_ts) {
+      ncols <- 4
       col_types <- paste(c("c", rep("d", ncols - 1)), collapse = "")
     } else {
+      ncols <- 3
       col_types <- paste(rep("d", ncols), collapse = "")
     }
-
 
     if (header) {
       dat <-